+++ axum-macros/tests/debug_handler/pass/multiple_extractors.rs
@@ -1,6 +1,7 @@
 use axum_macros::debug_handler;
+use axum::http::{Method, Uri};
 
 #[debug_handler]
-async fn handler(_one: String, _two: String, _three: String) {}
+async fn handler(_one: Method, _two: Uri, _three: String) {}
 
 fn main() {}