+++ axum-macros/tests/from_request/fail/enum_from_request_ident_in_variant.rs
@@ -1,0 +1,12 @@
+use axum_macros::FromRequest;
+
+#[derive(FromRequest, Clone)]
+#[from_request(via(axum::Extension))]
+enum Extractor {
+    Foo {
+        #[from_request(via(axum::Extension))]
+        foo: (),
+    }
+}
+
+fn main() {}