+++ axum-macros/tests/from_request/fail/enum_from_request_on_variant.rs
@@ -1,0 +1,10 @@
+use axum_macros::FromRequest;
+
+#[derive(FromRequest, Clone)]
+#[from_request(via(axum::Extension))]
+enum Extractor {
+    #[from_request(via(axum::Extension))]
+    Foo,
+}
+
+fn main() {}