+++ axum-macros/tests/from_request/fail/enum_no_via.rs
@@ -1,0 +1,6 @@
+use axum_macros::FromRequest;
+
+#[derive(FromRequest, Clone)]
+enum Extractor {}
+
+fn main() {}