+++ axum/src/routing/mod.rs
@@ -504,7 +504,7 @@ where
                 )
             }
             Err(MatchError::ExtraTrailingSlash) => {
-                let new_uri = replace_trailing_slash(req.uri(), &path.strip_suffix('/').unwrap());
+                let new_uri = replace_trailing_slash(req.uri(), path.strip_suffix('/').unwrap());
 
                 RouteFuture::from_response(
                     Redirect::permanent(&new_uri.to_string()).into_response(),
@@ -519,7 +519,7 @@ where
 }
 
 fn replace_trailing_slash(uri: &Uri, new_path: &str) -> Uri {
-    let mut new_path_and_query = new_path.to_string();
+    let mut new_path_and_query = new_path.to_owned();
     if let Some(query) = uri.query() {
         new_path_and_query.push('?');
         new_path_and_query.push_str(query);