+++ axum-core/CHANGELOG.md
@@ -9,6 +9,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - None.
 
+# 0.2.5 (08. June, 2022)
+
+- **added:** Automatically handle `http_body::LengthLimitError` in `FailedToBufferBody` and map
+  such errors to `413 Payload Too Large` ([#1048])
+- **fixed:** Use `impl IntoResponse` less in docs ([#1049])
+
+[#1048]: https://github.com/tokio-rs/axum/pull/1048
+[#1049]: https://github.com/tokio-rs/axum/pull/1049
+
+# 0.2.4 (02. May, 2022)
+
+- **added:** Implement `IntoResponse` and `IntoResponseParts` for `http::Extensions` ([#975])
+- **added:** Implement `IntoResponse` for `(http::response::Parts, impl IntoResponse)` ([#950])
+- **added:** Implement `IntoResponse` for `(http::response::Response<()>, impl IntoResponse)` ([#950])
+- **added:** Implement `IntoResponse for (Parts | Request<()>, $(impl IntoResponseParts)+, impl IntoResponse)` ([#980])
+
+[#950]: https://github.com/tokio-rs/axum/pull/950
+[#975]: https://github.com/tokio-rs/axum/pull/975
+[#980]: https://github.com/tokio-rs/axum/pull/980
+
+# 0.2.3 (25. April, 2022)
+
+- **added:** Add `response::ErrorResponse` and `response::Result` for
+  `IntoResponse`-based error handling ([#921])
+
+[#921]: https://github.com/tokio-rs/axum/pull/921 
+
 # 0.2.2 (19. April, 2022)
 
 - **added:** Add `AppendHeaders` for appending headers to a response rather than overriding them ([#927])